Allow persistent sessions
New functionality to allow persistent sessions has been added. This
includes the ability to store the session id as an encrypted cookie
and updates to the controller to check this and produce the correct
view.

The relevant test coverage has also been added.

Finally a small database migration has been added to include the
session id.